fix(news): adapt to BigDecimal price in Product entity

+++ src/main/java/iuh/wwwprogramming/service/impl/NewsServiceImpl.java
@@ -312,7 +312,7 @@ public class NewsServiceImpl implements NewsService {
 
         return products.stream().map(p -> {
             int discount = (p.getDiscount() != null) ? p.getDiscount() : 0;
-            Double price = (p.getPrice() != null) ? p.getPrice() : 0.0;
+            Double price = (p.getPrice() != null) ? p.getPrice().doubleValue() : 0.0;
             Double origPrice = (discount > 0 && price > 0)
                     ? (double) Math.round(price / (1.0 - (discount / 100.0)))
                     : null;
